\name{GeneNet-internal}
\alias{myrmvnorm}
\alias{ggm.list.edges}



\title{Internal GeneNet Functions}
\description{
  Internal GeneNet functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
