\encoding{latin1}
\name{GeneNet-package}
\alias{GeneNet-package}
\docType{package}
\title{The GeneNet package}

\description{GeneNet is a package for analyzing gene expression
  (time series) data with focus on the inference of gene networks.
  In particular, GeneNet implements the methods of Sch\"afer and 
  Strimmer (2005a,b,c) and Opgen-Rhein and Strimmer (2006, 2007)
  for learning large-scale gene association networks (including
  assignment of putative directions).  
}

\author{Juliane Sch\"afer, Rainer Opgen-Rhein, and Korbinian Strimmer (\url{http://strimmerlab.org/})}
\references{
  See website: \url{http://strimmerlab.org/software/GeneNet/}
}
\keyword{multivariate}
\seealso{
\code{\link{ggm.estimate.pcor}, \link{network.test.edges}, \link{extract.network}, \link{network.make.dot}.}
}
